Audio fingerprinting is now working end-to-end
Unfortunately, the `get_fingerprint_raw` method in the chromaprint library is unsound. In particular,
the lifetimes are not correct: the `Fingerprint` struct (owned) is dropped even though the inner `Raw`
has the lifetime of the context. This leads to a use-after-free bug. I'll need to dig into how this is
addressed in other FFI libs...

For the time being, we're using the FFI hash function directly.