Use resampler delay to correct fingerprint timestamps
This needs to be done to correct skew between input and output sample rates. In particular,
if the output rate is not a multiple of the input, then samples will be lost over time, causing skew.
To correct this, you can use the delay returned by the resampler to determine how much the current set
of samples need to be delayed to be presented correctly.

See: http://www.ffmpeg.org/doxygen/2.8/group__lswr.html#ga5121a5a7890a2d23b72dc871dd0ebb06