Fully working implementation!
The key fix was to flush the resampler whenever a delay is returned.
This is because the resampler will enqueue any samples that could not fit
in the output buffer. With this fix, the timestamp issue was also resolved
since we are now actually processing all audio samples instead of throwing
some out every frame.

Other than this, I made a bunch of other changes throughout.