Handle case of input parameters changing mid-audio stream
The ffmpeg software resampler will return `AVERROR_INPUT_CHANGED` if the input
frame has different parameters than the initial frame fed to the resampler context.

Instead of giving up on decoding the audio stream, we will initialize a new resampler
context and swap that out for the current context. If the stream parameters are not
changing often, this should have minimal impact on analysis/decoding performance.