Run search across all provided video pairs
To find the opening and ending for a given video, we first run a search for
the video and all other provided videos. Then, based on the results, we find
the longest/best opening and ending across all pairs and use that as the final
result to print out and/or save to disk.