Implement an LCS-based search for common frame hash strings
This runs in about O(n * m) time if you ignore the binary heap insertions.

Note that this about 2x slower than the original approach. However, it turns out
that the original approach was missing a few edge cases, which means that the
search was not actually work correctly... So, this commit kills two birds with one
stone.