feat(comparator): New match search logic
Prior to this change, finding the best match simply involved looking for the
_longest_ match among all other videos. This works most of the time, but breaks
down in cases where the longest match is, for example, a montage scene with no dialogue.

The new match logic considers two things:

(1) The number of _distinct_ matches for a particular sequence across all videos
(2) The sequence length

In other words, the logic now looks for sequences that are present in multiple videos.

At a high level, the new match logic works as follows:

1. Build a list of candidate sequences across both openings and endings.
2. For each candidate sequence, compare the sequence hash[1] with each other candidate.
3. If the sequence hash matches, we group these two sequences together as a match (using a set).
4. Build a list of distinct sequences sorted by the number of matches followed by the sequence length (one for openings
   and one for endings).
5. Take the highest ranked opening and ending from each list.